Backend/requirements.txt
flask
flask-cors
pandas
pillow
torch
torchvision
ftfy
regex
git+https://github.com/openai/CLIP.git